refactor: src 以下のディレクトリ構成を変更
